DecisionMakingUtils.jl

Documentation for DecisionMakingUtils.jl

DecisionMakingUtils.FourierBasis — Type

FourierBasis([::Type,] num_inputs::Int, dorder::Int, iorder::Int [, full::Bool=false])

Creates a struct to generate fourier features up to a given order. Both coupled fourier features, e.g., $\cos(π(3x₁ + 2x₂))$ and uncoupled features, $[\cos(πx₁), \cos(π2x₁), …]$ can be generated with this basis function. The dorder parameter controls order for the coupling features. The number of coupled features generated, $(\text{dorder}+1)^{\text{num\_inputs}}$, grows exponentially with dorder, so it is not reccomended for use with high deminsional vectors. The iorder parameter controls the order of the independent features generated. The full parameter determines if both $\sin$ and $\cos$ features are generated, if false only cos features are generated.

See also: FourierBasisBuffer

Examples

julia> f = FourierBasis(2, 1, 2);

julia> x = [0.0, 0.5];

julia> feats = f(x)
6-element Array{Float64,1}:
  1.0
  6.123233995736766e-17
  1.0
  6.123233995736766e-17
  1.0
 -1.0

DecisionMakingUtils.FourierBasisBuffer — Type

FourierBasisBuffer(ϕ::FourierBasis)

Creates preallocated buffers for a fourier basis to use to avoid making allocations on every call of the basis.

See also: FourierBasis

Examples

```jldoctest julia> f = FourierBasis(2, 1, 2);

julia> buff = FourierBasisBuffer(f);

julia> x = [0.0, 0.5];

julia> feats = f(buff, x) 6-element Array{Float64,1}: 1.0 6.123233995736766e-17 1.0 6.123233995736766e-17 1.0 -1.0

DecisionMakingUtils.GaussianNormalization — Type

GaussianNormalization(num_features::Int[, weight])

This is a functor that normalizes each element of a vector to be centered around a mean and have a variance of one. OnlineStats.KahanVariance is used for computing and tracking the mean and variance. Likewise any OnlineStats.Weight can be used to weigh each sample.

See also: ZeroOneNormalization, PosNegNormalization

Examples

julia> nrm = GaussianNormalization(2, 1e-4);

julia> x = [1.0, 0.0];

julia> feats = nrm(x)
2-element Array{Float64,1}:
 1.0
 0.0

julia> y = zero(x);  # create buffer to prevent allocations

julia> feats = nrm(y, x);  # no allocation return

julia> feats = nrm(y, x, fit=true)  # update mean and variance using x
2-element Array{Float64,1}:
 0.9999
 0.0

DecisionMakingUtils.PosNegNormalization — Type

PosNegNormalization(low::T, high::T)

This is a functor that normalizes a vector to be in the range [-1,1]. Initial upper and lower bounds for each element are needed.

See also: ZeroOneNormalization, GaussianNormalization

Examples

julia> low = [0.0, -1.0];

julia> high = [3.0, 0.5];

julia> nrm = PosNegNormalization(low, high);

julia> x = [1.0, 0.0];

julia> feats = nrm(x)
2-element Array{Float64,1}:
 -0.33333333333333337
  0.33333333333333326

julia> y = zero(x);  # create buffer to prevent allocations

julia> feats = nrm(y, x);  # no allocation return

julia> feats = nrm(y, x, fit=true);  # update upper and lower bounds if x is outside the range

DecisionMakingUtils.ZeroOneNormalization — Type

ZeroOneNormalization(low::T, high::T)

This is a functor that normalizes a vector to be in the range [0,1]. Initial upper and lower bounds for each element are needed.

See also: PosNegNormalization, GaussianNormalization

Examples

julia> low = [0.0, -1.0];

julia> high = [3.0, 0.5];

julia> nrm = ZeroOneNormalization(low, high);

julia> x = [1.0, 0.0];

julia> feats = nrm(x)
2-element Array{Float64,1}:
 0.3333333333333333
 0.6666666666666666

julia> y = zero(x);  # create buffer to prevent allocations

julia> feats = nrm(y, x);  # no allocation return

julia> feats = nrm(y, x, fit=true);  # update upper and lower bounds if x is outside the range

Base.length — Method

length(ϕ::FourierBasis)

Returns the number of feautes produced by the Fourier basis.